Move the ErrorAlert struct inside the AppState
