+++ Tensai/Views/TriviaQuizConfigView.swift
@@ -54,41 +54,8 @@ struct TriviaQuizConfigView: View {
     }
     
     // -------------------------------------------------------------------------
-    // MARK:- Private methods
+    // MARK:- Private method
     // -------------------------------------------------------------------------
-    
-    /// Presents an alert that displays an **Invalid Category** error.
-    private func presentInvalidCategoryErrorAlert() {
-        appState.errorAlert = ErrorAlert(
-            title: "Invalid Category",
-            message: "The category is no longer valid. Please try again."
-        )
-    }
-    
-    /// Presents an alert that displays a **No Results** error.
-    private func presentNoResultsErrorAlert() {
-        appState.errorAlert = ErrorAlert(
-            title: "Not Enough Questions",
-            message: "There are not that many questions in the database to "
-                + "start the quiz. Please try again."
-        )
-    }
-    
-    /// Presents an alert that displays a **Request Timed Out** error.
-    private func presentRequestTimedOutErrorAlert() {
-        appState.errorAlert = ErrorAlert(
-            title: "Could Not Start the Quiz",
-            message: "The request timed out. Please try again."
-        )
-    }
-    
-    /// Presents an alert that displays an “unknown error.”
-    private func presentUnknownErrorAlert() {
-        appState.errorAlert = ErrorAlert(
-            title: "Could Not Start the Quiz",
-            message: "An unknown error has occurred."
-        )
-    }
     
     /// Creates and starts a new quiz based on the customization settings.
     private func startQuiz() {
@@ -100,15 +67,15 @@ struct TriviaQuizConfigView: View {
             case .success(let response):
                 DispatchQueue.main.async {
                     if response.code == 1 {
-                        self.presentNoResultsErrorAlert()
+                        self.appState.errorAlert = .noResultsErrorAlert
                         return
                     }
                     if response.code == 2 {
-                        self.presentInvalidCategoryErrorAlert()
+                        self.appState.errorAlert = .invalidCategoryErrorAlert
                         return
                     }
                     if response.code > 0 {
-                        self.presentUnknownErrorAlert()
+                        self.appState.errorAlert = .unknownErrorAlert
                         return
                     }
                     UserDefaults.standard.set(
@@ -126,9 +93,9 @@ struct TriviaQuizConfigView: View {
                     }
                 }
             case .failure(.requestTimedOut):
-                self.presentRequestTimedOutErrorAlert()
+                self.appState.errorAlert = .requestTimedOutErrorAlert
             default:
-                self.presentUnknownErrorAlert()
+                self.appState.errorAlert = .unknownErrorAlert
             }
         }
     }
